remove six used
